# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Church check-in and registration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkins/batch:
    post:
      operationId: batchCheckin
      tags: [checkins]
      summary: Check in multiple children with a shared family security code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchCheckinInput"
      responses:
        "201":
          description: Batch check-in result with shared labels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchCheckinResult"
  /checkins:
    get:
      operationId: listCheckins
      tags: [checkins]
      summary: List today's check-ins
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: formId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckIn"
    post:
      operationId: createCheckin
      tags: [checkins]
      summary: Check in a child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "201":
          description: Check-in record with label data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInWithLabel"
        "409":
          description: Child is already checked in (active check-in exists)
          content:
            application/json:
              schema:
                type: object
                required: [error, checkinId]
                properties:
                  error:
                    type: string
                  checkinId:
                    type: integer
  /checkins/{checkinId}:
    patch:
      operationId: updateCheckin
      tags: [checkins]
      summary: Update a check-in record (e.g. add notes)
      parameters:
        - name: checkinId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCheckinInput"
      responses:
        "200":
          description: Updated check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
    delete:
      operationId: deleteCheckin
      tags: [checkins]
      summary: Delete a check-in record (undo an accidental check-in)
      parameters:
        - name: checkinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /checkins/{checkinId}/checkout:
    put:
      operationId: checkoutChild
      tags: [checkins]
      summary: Check out a child
      parameters:
        - name: checkinId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Updated check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
  /checkins/{checkinId}/undo-checkout:
    put:
      operationId: undoCheckout
      tags: [checkins]
      summary: Undo a check-out (clear checkoutAt, restoring child to checked-in state)
      parameters:
        - name: checkinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckIn"
  /event-categories:
    get:
      operationId: listEventCategories
      tags: [eventCategories]
      summary: List all event categories
      responses:
        "200":
          description: List of event categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventCategory"
    post:
      operationId: createEventCategory
      tags: [eventCategories]
      summary: Create a new event category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCategory"
  /event-categories/{categoryId}:
    patch:
      operationId: updateEventCategory
      tags: [eventCategories]
      summary: Rename an event category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventCategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCategory"
    delete:
      operationId: deleteEventCategory
      tags: [eventCategories]
      summary: Delete an event category (reassigns its events to General / Other)
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted
...
components:
  schemas:
...
    CheckIn:
      type: object
      required: [id, childId, childFirstName, childLastName, checkinAt]
      properties:
        id:
          type: integer
        childId:
          type: integer
        childFirstName:
          type: string
        childLastName:
          type: string
        guardianName:
          type: string
        checkinAt:
          type: string
        checkoutAt:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
        labelCode:
          type: string
        labelPrinted:
          type: boolean
        pickupPersonName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        updatedAt:
          type: string
    CheckInInput:
      type: object
      required: [registrationId]
      properties:
        registrationId:
          type: integer
        room:
          type: string
        sessionId:
          type: integer
    CheckoutInput:
      type: object
      properties:
        pickupPersonName:
          type: string
        notes:
          type: string
    UpdateCheckinInput:
      type: object
      properties:
        notes:
          type: string
    CheckInWithLabel:
      type: object
      required: [checkin, labelData]
      properties:
        checkin:
          $ref: "#/components/schemas/CheckIn"
        labelData:
          $ref: "#/components/schemas/LabelData"
    LabelData:
      type: object
      required: [childName, labelCode, checkinDate]
      properties:
        childName:
          type: string
        guardianName:
          type: string
        labelCode:
          type: string
        checkinDate:
          type: string
        room:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        specialNeeds:
          type: ["string", "null"]
        organizationName:
          type: string
...
    BatchCheckinInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [registrationId]
            properties:
              registrationId:
                type: integer
              room:
                type: string
    BatchCheckinResult:
      type: object
      required: [checkins, labels]
      properties:
        checkins:
          type: array
          items:
            $ref: "#/components/schemas/CheckIn"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/LabelData"
...
    EventCategory:
      type: object
      required: [id, slug, name, isDefault, createdAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
    CreateEventCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateEventCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string