# lib/api-zod/src/generated/types/checkInInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church check-in and registration API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInInput {
  registrationId: number;
  room?: string;
  sessionId?: number;
}