# lib/api-zod/src/generated/types/eventCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church check-in and registration API
 * OpenAPI spec version: 0.1.0
 */

export interface EventCategory {
  id: number;
  slug: string;
  name: string;
  isDefault: boolean;
  createdAt: string;
}